[backend] added orders table
